Fixed glitch in path-bar margin
Previoulsy the 1px margin in between inner button in path-bar was sometimes
appearing also between the first slider-button and the first text-button,
where is not expected, and then disappearing after hovering on the latter.

This commit simplifly the logic to set this margin only on inner path-bar
widgets and that glitch is no more visible.