Fix focus ring horizontal widget in headerbar
Normally, horizontally linked widgets share right/left border, so
when the widget on the left if highlighted, also the left border of
the sibling on the right is highlighted as well.

This commit removes the logic above in headerbars, where horizontally
linked widgets are forcibly separated and do not share any border.

closes #68