Only upload to the store PR and push to master
We want people forking the repo being able to try the snap.
For that, only PR against the main repo ("ubuntu" owner namespace)
or push event to master will be taken into account.

The others will just build the snap.
So, a PR from a branch in the ubuntu snap will build and release from the PR
job, and only build from the corresponding branch build job.