Fixed headerbar separator color
separator in a contextual menu of pathbar (e.g. Nautilus) takes the same
color of a headerbar separator. Being the latter dark and the first
bright, the same color can not be used, causing issue #576

Reorganized the code to avoid to style such contextual menu separator as
if it was in a headerbar.

closes #576