Style circular button in Nautilus
Circular button is used in Nautilus as as stop button while Nautilus is
loading a folder. With new dark colors for the "loading" label the
button was barely visible.

Added brighter button background, some margin and a red hover color
(stop loading is not strongly destructive, but will prevent the folders
to be completely visible until refresh, so I think red color suits fine).

closes #705