drawing/entries dark: fixed backdrop entry backdround color
In backdrop, dark entries background got darker by just 1%, however this
little change was enough to make the background the same as transparent,
loosing the difference from insensitive-backdround entries.

To fix it, this commit remove the background color change in backdrop
for dark entries, as keeping this same color in both normal and backdrop
state does seems to work fine as well.