gtk2 assets/line: Adjust line asset transparency
line asset is used as separator with both light and dark background
(e.g. in file chooser folder separator and as menu item separator),
for this reason the dark color, set in the previous commit, with full
opacity can not fit both usages.

This commit adjusts the transparency of the line color to let the asset
be usable in both use cases, being visible on dark backgrounds and
not too dark on light ones.