common/scale circular slider: made border transparent in dark variant
In the dark variant, gtkscale slider border looks fuzzy, probably due to
the different contrast between a dark background (in the light variant
the background is brighter and the border does not look fuzzy).

With this commit, the gtkscale slider border is set to transparent to
dark variant only to have a sharper border.

NOTE: a similar change is to be made to the 'scale with marks', however
that is a svg image and belongs to the gtk/dark folder which has not yet
landed on master.

see #855