Fix transparent scale in jet popover
Jet popover uses a wildcard to set transparency for all widgets, however
the elements of the scale widget, that is slider, trough and highlight do
not require to be transparent or are invisible otherwise.

This change limit the wildcard for these elements.
NOTE: this same fix can be used if other widgets present the same issue

closes #1067