Allow generate-symlink providing a pattern
`generate-symlink` now accepts one argument to filter out
any `.list` entry that does not match with such argument.
With this change, it is possible to generate only specific symlinks.

If no argument is provided, the script falls back to the usual behavior,
symlinking all the entries in any `.list` file.

e.g.

  the command

    $ ./generate-simlink.sh folder-open

  generates only symlinks for `.list` entries that concern
`folder-open`, both as source and destination

  folder-open folder          (OK)
  folder-close folder-open    (OK)
  folder-one   folder-two     (NO)
  image-x      image-y        (NO)