Build snaps only for bionic branch
As per Gnome shell changes on Disco, we are going to freeze Bionic
development. For this reason, any future change for the snap package
will be build only if pushed on the dedicated `bionic` branch.
Moreover, the majority of PRs will be for master branch, so there is no
real need to build testing snaps for each of them.

Summing it up:
- Stop building testing snap for each PR
- Build and push on edge channel only merge on `bionic` channel