Fix path-bar-box with width-maximized class
Latest nautilus added a `.width-maximized` class style that is applied
to path-bar when the windows is "big enough". In such situation, the
path-bar changes from left-aligned to centered and it starts moving
together with the `search` button. Yaru did not show any connection
between path-bar and search icon, so the two widgets looked floating
in the headerbar with no reason.

Add a path-bar border (only when `.width-maximized` class exists) that
starts from the right end towards the search button to show that a
connection exists (very much like Adwaita does).

Closes #1304