tweaks: darker headerbar variation
headerbar background color base #424040 might be too bright. This change
rebases the "lighter headerbar" on a 6% darker color #323030, which is
still 3% brighter than the previous one. To keep the visibility of the
checked button borders, the checked button background has been made 3%
brighter as well, keeping a very dark absolute color.

Considering the other guidances (light effect on buttons), this dark
headerbar should allow the buttons to be visibile anyway.