build: fix ppa build on bionic
automatic deb build on bionic is broken since we introduced build of both light and dark shell variant.
This is probably due to meson version, since in eoan it is not reproducible, however the error is that meson is not able to create different target lists for light and dark variants builds.

gnome-shell/src/dark/meson.build:25:2: ERROR: Tried to create target "style-gnome-shell-high-contrast", but a target of that name already exists.

This commit changes the target name for dark variant shell build keeping the output filename the same. This way meson does not complain and shell switch works.