shell: fix shell installation directories
Currently yaru gnome-shell theme is installed in the themes directory
/usr/share/themes/<themenane>/gnome-shell, which is where User-Themes
extension looks for 3rd party themes and a symlink is created under
/usr/share/gnome-shell/theme, where Gnome-shell looks for the session's
default theme.

Other than not being formally correct (the default theme should go into
the gnome-shell session's folder), this creates a dependency of the
gnome-shell-theme debian package from the gtk-theme package, that
creates the <themes>/Yaru folder first.

To fix the form and void adding dependency between packages (ideally,
gnome-shell could be installed without the gtk part), Yaru will be
installed under gnome-shell directory and the symlink will be created in
the themes directory, together with the installation of and Yaru-dark
shell, which is the alternative.