apps: increase difference between gnome-weather buttons
Currently it is hard to differentiate between checked and unchecked
buttons in gnome weather, since the (dark) colors are very similar. In a
previous Yaru version, this app was customized for this very same
reason.

Adding back button customization in apps scss file, using the same logic
Yaru dark buttons have (darker means checked, lighter means unchecked),
forcing the colors a little bit, since the bright color around the
buttons (if the weather is clear, the background is blue sky).

closes LP:1807508