gtk: notebook's tab is an extension of the notebook
This change re-apply previous Yaru style for tab in a way that they
look like an extension of the notebook they belong to.

NOTE: here "overshoot" is used also as undershoot.
The undershoot is used to style the effect of the notebook's content
sliding below the tab and was removed from upstream