Dark theme: darken entries in notebooks
Since notebooks and entries both use $base_color, entries drown in the dark theme, looking almost disabled. Darkening them for the dark theme only in notebooks helps here. We reduce the change for notebooks only since it is safe to say that the $base_color won't be changed by apps. Everywhere else it is not clear where the entries are drawn.