remove dark/light suffix from gresource name
In light and dark gtk-3.20 meson.build, the gresource target was
modified for a meson complain about the name of the custom_target being
already present. This causes the gresource to have the wrong name once
installed, breaking the variants.
Trying first removing the suffixes (it is not clear why the
complain, since the builds are in different folders). If this doesn't
work, a post_install script to renaming the target can be attempted.