switch: stronger disabled state in default and light variants
In default and light variant, the disabled checked switch is the same as
non-disabled once. In the dark variant, the only difference is the
handle.

This is due to the need to move our switch style to tweaks to reduce
upstream diff in common, so the fix consists in porting some missing
styling from common to tweaks.

Closes: #2016