gnome-shell-theme.gresource: Use proper icons subdirectories
Gnome shell uses `gtk_icon_theme_add_resource_path()` to add the
resources 'icons' subfolder as a theme lookup location, however it just
puts the icons into the root of this folder, and as per definition this
location is just the "ultimate fallback" where to look for icons.

Now, when we load an icon such as the keyboard ones the actual GIcon
with fallbacks that we load includes:

Loading gicon with fallbacks for the 'keyboard-enter-symbolic' would cause the
actual GThemedIcon to include as fallback all these icons
 - 'keyboard-enter-symbolic', 'keyboard-symbolic', 'keyboard-enter', 'keyboard'

Since yaru inherits from Humanity, and given that in that theme we've an
icon under a category-path called just 'keyboard', then we'd simply load
that, because due to the category it will have an higer priority than ours.

To fix this, and to ensure that the gresource icons will always have the
priority, let's move our icons under a proper category subdirectory, in
order to increase their priority over other themes.