gnome-shell/calendar: darken event dot
In the default Yaru shell theme, the calendar dot of days having an
appointments is not visible. The dark shell theme is not affected.

The problem is due to the color of calendar-today.svg being too similar
to the calendar background.

Fix consists in changing the color of calendar-today.svg in a way that
is both visible in the light and dark shell theme.

Fixes: https://bugs.launchpad.net/ubuntu/+source/yaru-theme/+bug/1870938