shell/calendar: fix day color if event and today
With upstream fix, if calendar-today has events, the color is set to a
variant of fg_color, which is a dark color in our light variant, while
we need a light one, since the date is inside the selected orange
circle.

It is not necessary a check for light/dark variant, because the
selected_fg_color is set accordingly already and takes care of the
differences eventually.