Rework extract symbolic icons (#2303)
* Remove trailing whitespaces

* Rework get_output_filename

- use a (probably) clearer naming
- add function documentation

* use docopt to parse CLI inputs

To improve discovery of the script features (parse all the icons, or
just some of them), running the script without args is forbidden.

Use docopt to generate the appropriate CLI args parsing from the usage
doc.

The new arguments are:
- all: extract all the available icons (as it was before without passing
  any argument)
- only, followed by a list of icon names: extract only the given list of
  icons

* Add force rendering to CLI options

* Rework main function

Try to rework main in order to:

1. get all available icons from svg file, as it did previously
    when extracting all the files
2. extract the icons into proper svg, all or only the selected
    once, according to CLI args

* Do not specify binaries path

Which inkscape and svgo binary use shall be decided by the system

* Improved documentation

* Fix scalable/README.md

* Add dependencies checker

do not permit the script to start if the dependencies are not met

* Capture output and exit code of external tools

External tools like inkscape and svgo can fail, but the output was
automatically dismissed, and the exit code ignores.

- use Open3 to capture output and exit code of syscalls,
- only in case of failure the command and its output is returned,
    and the process terminates

* Better check_deps error message

* update inkscape call

inkscape >1.0 does not support --file and --without-gui anymore
and inkscape <1.0 seems already able to support it.

* Add chopSVG documentation