Do not store LP issues json for lpbugtracker
Currently, lpbugtracker.py script needs a JSON file with the last tracked
Launchpad bugs to differentiate new bugs from old ones, and uses HUB to
create new bugs that mirror the ones on Launchpad.

Using HUB however we can also query the list of tickets already created on
GitHub to mirror the ones on Launchpad, without the need to create and
maintain the JSON file.