metacity: Add title bar left and right edge gradient
This is required for the Yaru-MATE theme only, it has not effect on the Yaru-MATE-dark and Yaru-MATE-light themes.

Do to the mix of lighter window borders and darker title bar borders in the Yaru-MATE themes, it is required to blend the left/right edges of the title bar from the darker to light border colours.