icons: Have the Yaru icons inherit from Ambiant-MATE or Radiant-MATE icons
  * Yaru-MATE and Yaru-MATE-dark inherit from Ambiant-MATE
  * Yaru-MATE-light inherits from Radiant-MATE

For the most part, inheriting from Ambiant-MATE and Radiant-MATE provides fallback for icons used in the AppIndicators, not currently exposed via Yaru icons.

This is a stop gap measure to prevent visual regressions. It may be possible to create symlinks to Yaru icons to render the fallback to Ambiant-MATE and Radiant-MATE obsolete.