Do not pre-run inkscape
With inkscape 1.0 something changed in the way inkscape runs in
background, and it seems than pre-running the script, creating a
inkscape thread pool, communicating later the command to execute is not
working properly.
As Dirty Fix(R) this commit replaces the inkscape thread pool with a
direct call to inkscape with the full command for each rendering.
This is very likely slower, but it should allow users to continue
working with our cursors during a deeper refactory of this script.