GNOME Shell 42 theme sync (#3403)
* gnome-shell/screenshot: Use yaru color for screenshot red button

* gnome-shell: Sync missing non-yaru changes with upstream

* gnome-shell: Move icons to gnome-shell-icons.gresource file

This is done by upstream as well, so follow that providing two resource
files. GNOME Shell require patches in order to load the icon theme.

* gnome-shell: Update theme as per GNOME 42 changes

Synchronize the yaru code to follow upstream changes.
Some changes have been dropped as not easy to apply to the new codebase
but they're very limited.