meson: Remove uneeded symlinks and use build system to figure out files to use
Make meson to re-use default theme thdefinitions and assets when a theme
doesn't provide a specific version for it.

As per this we can avoid having multiple symlinks but just ensuring
that we always prefer theme-specific files, and in case they don't exist
we fallback to the same theme variant or to the default theme.

As per this, we can avoid lots of duplication and symlinks that are not
easiest to maintain.

One note on sassc: it will now include multiple paths from where to pick
a specific included resource.
Note that if "@include 'fooo'" is used in a sym-linked resource it will
try to load that from the parent folder of the sym-link target resides.
So, we still need to preserve some symlinks otherwise sassc will just
try to load everything from the target folder.

In case a file needs to be overridden by a theme, ensure that it's going
to be loaded (`strace` is your friend).