gtk/meson: Auto-generate gtk-themes for accent colors
Use some scss magic to create accent color themes without having us to
re-create or re-define all the themes.

We use an accent-colors scss file as entry point that may define the
yaru variables for accent colors and then we load the standard themes.

In case such variables are defined, the accents are set.