gnome-shell: Always use variant-named assets in all the variants
Since we now may have multiple variants of the same assets, we need to
be sure of what is what and do not assume that the dark theme variant is
the default one.

So let's wrap our 'url' function override so that it always include the
theme variant suffix on the colorable assets, and add it when generating
the theme.

This will make easier to switch from light to dark as default variant.