icons: Add meson targets to automatically render bitmaps
Instead of going through the manual scripts now it's possible to just
use meson to render the icons.

A new target called `render-icons` is provided, plus more for each
variant (such as `render-icons-default`) that can be run just using
`ninja` or `meson compile`.

In order to support this, now render-bitmaps supports input/output
folders as parameters.