icons: Replicate symlinks in installed folder when using meson >= 0.61
As per some relases, meson is doing copies of local symlinks when using
install_subdir, this is causing that lots of data is duplicated, causing
the icons package to be bigger than it should be.

New meson supports install_symlink feature that will enable to create
symlinks in install path, without duplicating icons.

So, to replicate the local structure, read the symlink status in source
dir using find and exclude all the symlinks from the installation.
Eventually, use install_symlink to replicate those in the install path.