Remove an issue causing ".login-dialog-prompt-entry"
This commit removes a part of the scss file, which causes the login entry field focus borders to not be themed correctly.

This part causes ".login-dialog" to be missing in front of ".login-dialog-prompt-entry:focus", which causes the theme to fall back to the default color. If that default color isn't the same as the currently selected Yaru theme color, they mismatch.

closes #4102