gtk-2.0: Use bark theme assets to provide wartybrown accent
We're too late to add a new theme accent for both noble and oracular, but
bark is close enough to wartybrown that we can use it as fallback for
the assets of very old Gtk2 applications. As they're close to unusable at
the moment without a proper theme.

At the same time use the warty color for main theme visual or text
items.

Helps with #4197

LP: #2107421