Fix Caps-lock-is-on color in light theme
When a Polkit dialog is open to ask for the password, if the
user activates the Caps Lock, a message appears under the text
entry notifying that it is active, to warn the user.

In dark theme, this message is painted in yellow, which is fine
because it gives a good contrast against the background; but in
litht theme, a slightly darker yellow is used, which has poor
contrast. This goes against the accessibility (A11Y) requirements
of having, at least, a 4.5/1 contrast ratio in text.

This patch fixes this by setting the WARNING text color to black.